requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
boto3==1.12.5
botocore==1.15.5
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-OAuth==0.12
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jwt==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pynamodb==4.3.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
urlopen==1.0.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.0.0
grpcio-tools==1.28.1
grpcio==1.28.1